requirements.txt
streamlit==1.39.0
tensorflow==2.17.0
opencv-python==4.10.0.84
numpy==1.26.4
pillow==10.4.0
matplotlib==3.9.2
gdown==5.2.0